TB052, written by me, published by Microchip Technology
This application note uses the MSSP module for the PIC16F872 in SMBus
Master Mode.  The PIC16F communicates to multiple thermal sensors over
the SMBus and displays the results on an LCD.  The code was written in
Assembly in a relocatable format utilizing the Linker script file to
segment an allocate the compiled code throughout the chip.